+++ sympy/matrices/determinant.py
@@ -322,7 +322,7 @@ def _adjugate(M, method="berkowitz"):
     ========
 
     cofactor_matrix
-    sympy.matrices.common.MatrixCommon.transpose
+    sympy.matrices.matrixbase.MatrixCommon.transpose
     """
 
     return M.cofactor_matrix(method=method).transpose()
